pyproject.toml
[tool.poetry]
name = "multimodalwebagent"
version = "0.1.0"
description = ""
authors = ["schauppi <46676413+schauppi@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "src", from = ".", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-dotenv = "^1.0.1"
pyautogen = "^0.2.17"
openai = "^1.13.3"
colorama = "^0.4.6"
tzlocal = "^5.2"
playwright = "^1.42.0"
pytest-playwright = "^0.4.4"
instructor = "^0.6.4"


flask = "^3.0.3"
flask-cors = "^4.0.1"
pillow = "^10.3.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
webagent = 'src.oai_agent.oai_agent:main'
